fix: use int64 instead of int (32b integers) to represent task IDs everywhere
We've recently noticed that all integers representing task IDs coming
from Algolia were wrongly typed as 32b integers. However, the Algolia
engine is returning JSON numbers which can go beyond the range of 32b: a
JSON number can be as big as 2^53.

For this reason, this commit is changing the type of all the taskID
variables as well as all function accepting task IDs from `int` to `int64`.

This change was automated, for the most part, with the two following sed
commands:

```
find . -name '*.go' -exec gsed -Ei 's/taskID( +)int\)/taskID\1int64\)/' {} \;
find . -name '*.go' -exec gsed -Ei 's/TaskID( +)int/TaskID\1int64/' {} \;
```

Fix #591

+++ algolia/analytics/client.go
@@ -65,6 +65,6 @@ func NewClientWithConfig(config Configuration) *Client {
 	}
 }
 
-func (c *Client) waitTaskSearchClient(index string, taskID int) error {
+func (c *Client) waitTaskSearchClient(index string, taskID int64) error {
 	return c.searchClient.InitIndex(index).WaitTask(taskID)
 }
